创建banner-item

+++ src/main/java/io/github/talelin/latticy/model/BannerItemDO.java
@@ -36,6 +36,6 @@ public class BannerItemDO {
     @TableLogic
     private Date deleteTime;
 
-    private Long bannerId;
+    private Integer bannerId;
 
 }
